conftest, == for int, remove arg defaults, remove meteo_location
